+++ src/app/search/search-filters/sockets.tsx
@@ -6,7 +6,7 @@ import {
   modSlotTags,
   modTypeTags,
 } from 'app/utils/item-utils';
-import { getSocketsByCategoryHash, isEnhancedPerk } from 'app/utils/socket-utils';
+import { countEnhancedPerks, getSocketsByCategoryHash } from 'app/utils/socket-utils';
 import { DestinyItemSubType, DestinyRecordState } from 'bungie-api-ts/destiny2';
 import craftingMementos from 'data/d2/crafting-mementos.json';
 import {
@@ -247,24 +247,38 @@ const socketFilters: FilterDefinition[] = [
     },
   },
   {
+    keywords: 'catalyst',
+    description: tl('Filter.Catalyst'),
+    format: ['query'],
+    destinyVersion: 2,
+    suggestions: ['complete', 'incomplete', 'missing'],
+    filter:
+      ({ filterValue }) =>
+      (item: DimItem) => {
+        if (!item.catalystInfo) {
+          return false;
+        }
+
+        switch (filterValue) {
+          case 'missing':
+            return !item.catalystInfo.unlocked;
+          case 'complete':
+            return item.catalystInfo.complete;
+          case 'incomplete':
+            return item.catalystInfo.unlocked && !item.catalystInfo.complete;
+          default:
+            return false;
+        }
+      },
+  },
+  {
     keywords: 'enhancedperk',
     description: tl('Filter.EnhancedPerk'),
     format: 'range',
     destinyVersion: 2,
     filter: ({ filterValue }) => {
       const compare = rangeStringToComparator(filterValue);
-      return (item: DimItem) => {
-        let enhancedPerkCount = 0;
-        if (item.sockets) {
-          // count enhanced perks
-          for (const socket of item.sockets.allSockets) {
-            if (socket.plugged && isEnhancedPerk(socket.plugged)) {
-              enhancedPerkCount++;
-            }
-          }
-        }
-        return compare(enhancedPerkCount);
-      };
+      return (item: DimItem) => item.sockets && compare(countEnhancedPerks(item.sockets));
     },
   },
 ];