+++ docs/CHANGELOG.md
@@ -1,5 +1,19 @@
 ## Next
 
+* Hid Solstice armor rerolling sockets from Loadout Optimizer too.
+
+## 7.27.0 <span class="changelog-date">(2022-07-24)</span>
+
+## 7.26.1 <span class="changelog-date">(2022-07-23)</span>
+
+* Added Solstice event challenges to the Progress page.
+
+## 7.26.0 <span class="changelog-date">(2022-07-17)</span>
+
+* Worked around a Bungie.net API bug where Vanguard reset count was reported under Strange Favor (Dares of Eternity) instead.
+
+## 7.25.0 <span class="changelog-date">(2022-07-10)</span>
+
 ## 7.24.0 <span class="changelog-date">(2022-07-03)</span>
 
 * Weapon perks now include community-sourced weapon and armor perk descriptions courtesy of [Clarity](https://d2clarity.page.link/websiteDIM) and [Pip1n's Destiny Data Compendium](https://docs.google.com/spreadsheets/d/1WaxvbLx7UoSZaBqdFr1u32F2uWVLo-CJunJB4nlGUE4/htmlview?pru=AAABe9E7ngw*TxEsfbPsk5ukmr0FbZfK8w#). These can be disabled in settings.
@@ -231,6 +245,7 @@
 * Loadout Optimizer Mod picker will now correctly update when switching between mod slots without closing Mod Picker.
 * Loadout Optimizer now correctly takes Echo of Persistence's class-specific stat reductions into account when generating sets.
 * The "Kinetic Slot" icon in Compare sheet now looks different from the "Kinetic Damage" icon.
+* Added `catalyst:` filter which accepts the following parameters `missing`, `complete`, and `incomplete`.
 
 ### Features